Lower coverage expectation to 95%
The coverage tools have improved and now also check for derived traits,
like Clone or Debug. Clone is required for Copy, so there's no reason to
explicitly test this. Debug is required for `assert_eq!` tests, but is
never executed, if the assert_eq succeeds.

In addition to that the `PartialOrd` for `SpecialVersion` is never
tested. But since this is just the default implementation, this should
be fine.